chameleon: Add Chameleon MIC port and set channel map for simple case
Add Chameleon MIC port which will be connected to a one-channel
microphone directly.
For simple sink type like internal microphone on Cros device, or Mic port on
Chameleon, the audio signal is over the air. We do not use link to
combine the source to the sink. We just set a default channel map to
them.

BUG=chromium:463813
TEST=not used yet
Change-Id: Ia3def1919a045c1e391268975e39dda8ef3a85b2
Reviewed-on: https://chromium-review.googlesource.com/269008
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Tom Tam <waihong@google.com>