faft: Deprecate the broken_warm_reset FAFT config
There are 3 boards using this broken_warm_reset FAFT config to work
around the warm reset problem. Parrot fix is moved to servod, check:
  https://chromium-review.googlesource.com/269138

Butterfly and Stout are not supported currently. If we add support of
them, the changes will go the servod, not FAFT.

So it is safe to deprecate the broken_warm_reset FAFT config.

BUG=chrome-os-partner:39744
TEST=None
CQ-DEPEND=CL:269138

Change-Id: Ib48056dccb1506140aba8d16a0a1c05fb0ed1810
Reviewed-on: https://chromium-review.googlesource.com/269139
Reviewed-by: Yusuf Mohsinally <mohsinally@chromium.org>
Commit-Queue: Tom Tam <waihong@google.com>
Tested-by: Tom Tam <waihong@google.com>