chameleon: Set default scale for Chameleon Mic port
We need to set a scale on Chameleon Mic port because the gain on
Chameleon is small. The recorded audio data is clear but the value is
small. This cause processing in audio_helper falsely judges the recorded
audio data as meaningless. Scale the recorded data up to avoid this
problem.

BUG=chromium:463813
TEST=run audio_AudioBasicInternalSpeaker test on Squawks.

Change-Id: I3cfda9ee2bf52a4b4e90143173e535d746426f2e
Reviewed-on: https://chromium-review.googlesource.com/269253
Reviewed-by: Tom Tam <waihong@google.com>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>