[autotest] Toggle display of 'Advanced Options' in 'Create Host' tab
The order of items on the Create Host tab has been reordered as
follows:

  Job Name
  Image URL/Build
  test selection
  advanced options
  control file
  host selector

The display of the advanced options can be toggled. The hostless
option under the advanced options panel is auto-selected if
the user chooses a test with name containing 'test_suites'.

BUG=chromium:483652
DEPLOY=afe
TEST=Tested by deploying changes to a moblab and verifying that
the advanced options display could be toggled and that the
hostless checkbox was auto-selected as desired.

Change-Id: I773c9189903e06b519f42a85464bdd2dc48554d2
Reviewed-on: https://chromium-review.googlesource.com/269259
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Matthew Sartori <msartori@chromium.org>
Tested-by: Matthew Sartori <msartori@chromium.org>