balance_pools.py: Support auto-balancing all boards at once.
Right now, deputies need to manually list all the boards they want to
rebalance. This is a pain for deputies as you need to go copy/paste the
list of boards from the email. Teach balance_pools.py to be capable of
auto-balancing all boards at once.

This is a step towards enabling auto-balancing pools without deputy
intervention (we just need to add heuristics that clarify when we
should auto-balance and when we should not).

BUG=chromium:485403
TEST=Run the script and autobalance all DUTs.

Change-Id: Ifc8945a25d3f4eb9a5d2afd914b80a97befd6ba5
Reviewed-on: https://chromium-review.googlesource.com/269851
Trybot-Ready: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>