[autotest] In lab inventory, ignore DUTs with no board.
From time to time (twice so far), we've had cases of hosts being
added to the database with a pool: label, but no board: label.
That caused failures when gathering inventory, because
status_history.HostJobHistory assumed (without particular
justification) that every board has board: and pool: labels,
and nowhere was there a check to deal with it.

This changes the code so that the host_board and host_pool
properties in HostJobHistory will return `None` when there's no
appropriate label, and changes lab inventory gathering to skip hosts
with no board label.

BUG=chromium:480026
TEST=Unit tests

Change-Id: Ic1065a34ba0e4a2d2019bf8ea89507b277499af6
Reviewed-on: https://chromium-review.googlesource.com/269609
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Mungyung Ryu <mkryu@google.com>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>