balance_pools.py: Only print changes by default.
Currently we print a lot of text for each change in balance_pools.py. This
leads to verbose output that is hard to read -- it's hard to see what
the errors are when there is a wall of unrelated text.

BUG=chromium:485403
TEST=Balance all pools.

Change-Id: I236982c1a71fcb083fe3c97967e12ad6643d5ced
Reviewed-on: https://chromium-review.googlesource.com/269853
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>
Trybot-Ready: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>