faft: Install Autotest dependency automatically
FAFT tests used to explicitly to install Autotest dependency when
booting into an USB image as it may not have Autotest there. It
makes a bit confusion if it is needed or not. This change removes
this confusion by always installing Autotest dependency when the
original connection to DUT is lost, meaning that a new RPC
connection has to be established, furthermore better to ensure
Autotest dependency is also there.

BUG=chrome-os-partner:16231
TEST=Removed the Autotest directory from the USB stick, ran
firmware_RecoveryButton test, and checked Autotest reinstalled.

Change-Id: I6cfc4f0bf3fb6b5386115a9464bb042059849bb5
Reviewed-on: https://chromium-review.googlesource.com/270101
Trybot-Ready: Tom Tam <waihong@google.com>
Tested-by: Tom Tam <waihong@google.com>
Reviewed-by: Yusuf Mohsinally <mohsinally@chromium.org>
Commit-Queue: Tom Tam <waihong@google.com>