[moblab] Increase minimum tick seconds to reduce load.
This increases the minimum number of seconds for a given tick.
This reduces CPU load and apache load on MobLab but will impact
individual test times, however we Moblab can accept this hit
as it is not running tests on a time critical path.

BUG=chromium:481301
TEST=None

Change-Id: I5937f256375137acc5571a52eeaa7cadda7b6d04
Reviewed-on: https://chromium-review.googlesource.com/270474
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>
Reviewed-by: Laurence Goodby <lgoodby@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>