sequences: schedule jobs within a sequence job.
Instead of launching tests that last a week or more,
sequence a large number of smaller tests (few hours).

This way, if moblab crash, the current test will be marked as failed,
but the tests in the suite will be scheduled.

Note: retention test is still 7 days long.

BUG=chromium:372952
TEST=Run storage_qual_suspend_quick.
Check the child job control files are correct,
check the AFE is taking the tests into account.

Change-Id: I6c276dba1e537e454d9098bc2e99209a871fec57
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/270789
Reviewed-by: Simran Basi <sbasi@chromium.org>