autotest (wifi): migrate network_WiFi_Regulatory to hostapd-2.3
For the network_WiFi_Regulatory test, we want to make sure
that the Capabilities Information in the AP's beacons and
probe responses indicates that spectrum management is
enabled.

Previously, we did this with a local patch to hostapd. With
hostapd 2.3, there is a configuration option to enable this.
Given this, we're not carrying our patch forward to 2.3

To maintain the same behavior for this test, we
a) teach HostapConfig about the new hostapd option, and
b) have network_WiFi_Regulatory enable this option.

While there:
- Have HostapConfig._get_default_config() use an
  OrderedDict. This is necessary because the new hostapd
  option depends on other options to be configured first.
  (Order matters in the hostapd config file.)
- Add a packet capture to the test. (This was needed to
  verify that the right bit was set OTA. And it seems
  worthwhile to commit the change.)

Note that behavior for other tests will change after we
update to hostapd 2.3. Namely, other tests will send
beacons and probe responses with the Spectrum Management
set to false. I believe this is okay. (I believe other
tests do not require the Spectrum Management bit to be
set to true.)

BUG=chromium:487317
TEST=network_WiFi_Regulatory (on peppy, jerry, speedy, candy)
TEST=manual
CQ-DEPEND=CL:272531

Manual test: open pcap in wireshark, find a beacon frame,
and verify that "management frame -> Fixed parameters
-> Capabilities Information -> Spectrum Management" is
set to 1.

Change-Id: I6dcfeea8b174a2f0324b632249bd7aadce3c4a3d
Reviewed-on: https://chromium-review.googlesource.com/272536
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Zeping Qiu <zqiu@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>
Trybot-Ready: mukesh agrawal <quiche@chromium.org>