Autotest: Fix touch_WakeupSource flake.
On a small number of machines, for whatever reason, certain numbers are out of
sync.  Find the power/ folder for a device in a more reliable manner.

Also, add an exception for Rambi.  It should have the touchpad as a wake source,
but as it is not a shipping device that never happened for Atmel touchscreen
versions (see crbug.com/431003).

BUG=chromium:488300
TEST=ran this test on machines that previously failed and on some control
machines.

Change-Id: I3d7b2e61dcd2b365caf6d64505b659053e22d45e
Reviewed-on: https://chromium-review.googlesource.com/272638
Reviewed-by: Mussa Kiroga <mussa@chromium.org>
Reviewed-by: Kris Rambish <krisr@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Commit-Queue: Katherine Threlkeld <kathrelkeld@chromium.org>