Suspend: Reboot machine if connection lost
Add an option to reboot if connection to the autotest server is lost.
Also add a minimal amount of time to be up to ensure we have time to set
the link back.

BUG=chromium:418334
TEST=Check on a machine: hardware_StorageQualSuspendStress.test
when suspend/resume: No impact
Disconnect the ethernet cable from the machine for few minutes: check
the machine rebooted and test failed:
"Autotest client terminated unexpectedly: DUT rebooted during the test run."

Change-Id: I80eebeda90f05c0ea83aa218aacadcc86df4fe20
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/272699
Reviewed-by: Simran Basi <sbasi@chromium.org>