A test to help update chaos_dynamic_ap_list.conf.
We are going to start adding a lot of APs to the config file.
In order to reduce human error this test will fill in most of the
information and then the result can be pasted into the
chaos_dynamic_ap_list.conf file.

BUG=None
TEST=None

Change-Id: I5586641a6940b5197e8e86cbf1361ca94e244859
Reviewed-on: https://chromium-review.googlesource.com/175654
Commit-Queue: Kris Rambish <krisr@chromium.org>
Trybot-Ready: Kris Rambish <krisr@chromium.org>
Tested-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Tien Chang <tienchang@chromium.org>