[autotest] Provide a repair recommender for lab inventory.
This enhances the lab_inventory script to include a feature that
will recommend a list of hosts for repair.  The hosts are selected
so that they are (somewhat) close together, and all hosts are
guaranteed to reside in the same lab.  The selection is also
designed to try and improve the inventory numbers by focusing
on the boards that have the lowest number of working spares.

BUG=None
TEST=unit tests; various tests with --print; full run w/ e-mail

Change-Id: Ie9863ce337045a3fdd6d8d649613dda830fae7f0
Reviewed-on: https://chromium-review.googlesource.com/274541
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>