chaos_analyzer: Rework ACK validation
Fix a couple issues with ACK validation:

 - The attempt to copy "packet_iterator" was incorrect.
 - Only ACKs to DUT frames were checked.

The first issue is due to the fact that a trivial copy of
an iterator object yields a second reference to the same
iterator.  Thus, calling "copy_iterator.next()" also
advances packet_iterator.  This causes issues in captures
where the next packet in packet_iterator is not an ACK but
is a different packet that the state machine is looking for.

The second issue is due to the fact that the packet filter
did not admit ACKs directed at the AP.  These frames are
worth validating since it indicates whether the DUT heard the
associated packet.

In order to fix both issues, this CL reworks the ACK check
by going back to the packet capture file and retrieving the
very next frame in the packet capture and checking if it is
an ACK back to the transmitter of the current frame.  Doing
it this way avoids having to expand the packet capture
filter to admit extra traffic from the AP, but allows
individual packets outside of the filter to be validated.
Doing so also prevents the packet_iterator from being modified
during the test, nor does it incur the cost of a deep-copy
like itertools.tee().

BUG=chrome-os-partner:38849
TEST=Attempt 5 in chrome-os-partner:38849 yields
  Found state: WLAN AUTH REQUEST, Packet number: 79, Data rate: 6Mbps.
  WARNING! Missing ACK for state: WLAN AUTH REQUEST.
  Found state: WLAN AUTH RESPONSE, Packet number: 80, Data rate: 6Mbps.
Note that packet 80 is successfully decoded although it was also
evaluated to test whether it was an ACK to packet 79.

Change-Id: Ice8ab352335529e4ac919a2c97cfe2ad232a6fad
Reviewed-on: https://chromium-review.googlesource.com/275360
Reviewed-by: Roshan Pius <rpius@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>