chaos_analyzer: Add error details and packet warnings
Report the reason code and sender for deauthentication and
disassociation events.  This is accomplished by adding a
"details" field to the ErrorStateInfo tuple which reports
which fields of the packet might be interesting to include
in error messages.

Add more specific filters to the authentication response and
association response states so that only successful responses
are matched.

Add a list of packets for which warnings should be generated.
This uses the ErrorStateInfo named tuple, but does not terminate
the state machine.  Also it only matches packets that do not
occur naturally either in the history or future of the state
machine.  This makes it convenient to match the non-successful
versions of the authentication and association response frames.

BUG=chrome-os-partner:40901
TEST=Attempt 8 in chrome-os-partner:40901
WARNING! WLAN ASSOC REJECTED found, Packet number: 168, Status Code: 0x000c"
TEST=Attempt 3 in chrome-os-partner:40896
State Machine encountered error due to WLAN DEAUTH REQUEST, Packet number: 14, Sender: 24:f5:aa:9a:be:29, Reason Code: 0x0007"

Change-Id: Ie5bb93cd00b67cd2a9dc05ccb3afbb3ba725d726
Reviewed-on: https://chromium-review.googlesource.com/275315
Reviewed-by: Roshan Pius <rpius@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>