network_WiFi_UpdateRouter: update stable versions
Update the stable versions, to move the APs in the lab to
a build with hostapd-2.3. (The new version of hostapd resolves
a problem with the OverlappingBSSScan and bcm4354.)

BUG=chromium:492727
TEST=network_WiFi_UpdateRouter
TEST=suite:wifi_matfunc on a veyron_speedy

Note: The UpdateRouter test failed, but the failure was spurious.
See http://crbug.com/496868 for the details.

Note also: There were some failures in wifi_matfunc, but none
were regressions compared to the previous stable version. See
http://crbug.com/492940 for the details.

Change-Id: I566829adabde912886ce8a3d81396379e5cb717e
Reviewed-on: https://chromium-review.googlesource.com/275326
Trybot-Ready: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Kris Rambish <krisr@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>