autotest (wifi): improve shill_xmlrpc_server exit behavior
When shutting down the shill_xmlrpc_server, we attempt to start
the UI. That's reasonable, because we probably shut down the UI
when starting the shill_xmlrpc_server. (I believe we do this to
ensure that the UI does not, e.g., initiate WiFi scans.)

When starting the UI, we wait until the login prompt is visible.
This causes a couple of complications. First, in the normal case,
it slows down the exit of shill_xmlrpc_server by about 4 seconds.
Second, if the developer has switched the DUT to VT2, then the
login prompt will never become visible. Instead, we'll wait 90
seconds, before timing out the shut down of the shill_xmlrpc_server.

This second case is doubly unfortunate. First, we slow down
the running of the test. Second, we spam the test results with
errors that are othogonal to the purpose of the test.

To resolve this, we make starting the UI a non-blocking operation.
This does leave the possibility that the login prompt will not
be visible at the time of the next test. But I think that's ok.
If a test requires the UI to be visible, it should check that
condition itself. And, in any case, such a test will likely
emit a more relevant failure than a timeout in the shutdown of
shill_xmlrpc_server.

BUG=chromium:495837
TEST=network_WiFi_DisconnectClearsIP

Change-Id: I2ff2e4a35dee6dc6260d7abb7ef41aeda3d01afa
Reviewed-on: https://chromium-review.googlesource.com/275698
Trybot-Ready: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>