[autotest] Fix an error in an error report in build_externals.
For a certain error case in build_externals, the error report code
tried to create a string by saying <list>.join(<string>).  The
resulting exception meant that the original error failed to be
reported.

This fixes the call to the correct <string>.join(<list>)

BUG=None
TEST=None

Change-Id: I75e9badd72c9b41a6008b1886a7afcfbfaddba18
Reviewed-on: https://chromium-review.googlesource.com/277087
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Matthew Sartori <msartori@chromium.org>