autotest: add log messages for WiFi autotest start and finish
Add log messages on autotest server and DUT indicating when
1) WiFi autotest setup is complete, and the body of a test is
   about to start running, and
2) When the test is complete, and teardown is about to begin.

1) is logged at the end of wifi_cell_test_base.warmup(), since
warmup() is called right before the main body of the test, run_once(),
is executed. 2) is logged at the start of wifi_cell_test_base.cleanup(),
which marks the beginning of test teardown.

On the DUT, these log messages are added to the shill log
(/var/log/net.log) as debug messages. On the autotest server,
these messages will be found in the output of test_that.

BUG=chromium:492267
TEST=Run a WiFi autotest and look for the added test start and end
log messages in /var/log/net.log on the DUT, and in the test_that
output on the server.

Change-Id: I7da241b752ed22c5fc4f5734c7e98be539cfb186
Reviewed-on: https://chromium-review.googlesource.com/277130
Tested-by: Samuel Tan <samueltan@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Samuel Tan <samueltan@chromium.org>