Autotest: make touch_WakeupSource support HID touchpads.
Some new devices report whether their touchpads are a wake source differently.
Rather than a file called wakeup in the folder associated with the touchpad,
these devices report the wakeup status in the parent bus's folder.

Change the test to search for the parent bus and check for this file if it
cannot find a file for any of the inputs.

BUG=chromium:499903
TEST=ran on all existing combinations of touchpad/touchscreen wake

Change-Id: I19baab797f3ded96edbae4dc1aab8e09653affc3
Reviewed-on: https://chromium-review.googlesource.com/277151
Reviewed-by: Kris Rambish <krisr@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Commit-Queue: Katherine Threlkeld <kathrelkeld@chromium.org>