graphics_dEQP: add ability to run individual tests
The filter option previously allowed the running of individual tests,
but no longer does. We need this capability, so this patch adds
a test_names option that works with the filter_args control file.
If test_names is given, the filter option and expectation lists are
ignored. When test_names is not given, the filter option and
expectation lists work as before.

A test name can be a full name or just an initial part. If an
initial part is given, all tests that start with that part will
be run.  The option also allows multiple test names in the form
of a comma-separated list. Multiple names must be for the same API
(e.g., dEQP-GLES3, dEQP-GLES2, etc.).

BUG=None
TEST= test_that --args="test_names=dEQP-GLES3.info.version" \
             <dut_addr> graphics_dEQP.filter_args
	  test_that --args="test_names=dEQP-GLES3.accuracy,dEQP-GLES3.stress" \
             <dut_addr> graphics_dEQP.filter_args

Change-Id: I215bbef8a983884481c2ca9bad1643f943dadc3e
Signed-off-by: Mike Mason <michael.w.mason@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/278090
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Trybot-Ready: Frank Henigman <fjhenigman@chromium.org>
Commit-Queue: Ilja Friedel <ihf@chromium.org>
Trybot-Ready: Ilja Friedel <ihf@chromium.org>
Tested-by: Ilja Friedel <ihf@chromium.org>