Add automated test for FQDN
Some modems provide a Fully Qualified Domain Name option in their
ACK response.  It's been discovered that some versions of dhcpcd
can crash when trying to convert this to a D-Bus event.  This
test ensures that after a fix is in place that this problem does
not regress.

BUG=chromium:500112
TEST=Run this test

Change-Id: I82db1283bc2b289148b891bbeaa0eb276629e362
Reviewed-on: https://chromium-review.googlesource.com/280380
Reviewed-by: Rebecca Silberstein <silberst@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>