[autotest] Change lab status build check to use an anchored pattern.
When using the lab status to block a build using a regular expression,
this now uses re.match(), causing the search to be anchored at the
beginning (but not the end).  This makes it easier to do the right
thing when you want to block "quawks" but not "squawks", among other
cases.

BUG=None
TEST=run the (new) unit tests

Change-Id: I34f0f98df8c2a0506c5b56ff9d628683c610f7c0
Reviewed-on: https://chromium-review.googlesource.com/280840
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>