security_Firewall: Tentatively re-enable webservd tests
This partially re-enables testing of webservd firewall rules.
It does a number of things to reduce the incidence rate of
false failures:

  - Instead of just testing that webservd is installed, it
    also checks to see if the service is running.  If it
    is not (and does not appear in a short amount of time)
    the webservd rules are not checked.

  - If only the webservd rules are missing an error is
    printed but the test does not fail.

Both failures above should probably be considered as a
failure in webservd, not in the firewall test.  If folks
want to test proper function of webservd, presumably
a different test should be created.  Log messages for
both cases (webservd not running, missing webservd rules)
are printed, so incidence rates of these non-fatal checks
can be monitored.

While here, remove an errant reset of the "failed" variable
since everyone that looked at line believes it is here in
error.

BUG=chromium:470701
TEST=Run this test on an amd64-generic VM with:
TEST=(a) system at boot
TEST=(b) system after "stop webservd"
TEST=(c) system after "start webservd"
TEST=(d) system after "iptables -D INPUT 1" (removes 1st webservd rule) [FAIL]
TEST=(e) system after "iptables -D INPUT 1" (removes 2nd webservd rule)

Change-Id: I9ba1bba25559472d03ced34e96dc0c121b1924dd
Reviewed-on: https://chromium-review.googlesource.com/280889
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>