graphics_utils: fix get_external_resolution and get_external_crtc()
The original implementation assumes the existence of internal display
and its CRTC at index 0 of the CRTC array so the CRTC of external
display starts at index 1. This commit relaxes the assumption and
treats the CRTC at index 0 as an external one if the internal display
is not present.

BUG=chromium:429932
TEST=ran suite:chameleon_hdmi on veyron_mickey and link.

Change-Id: I1f8b9f120bbc8bc48a4aeea904889967c7206deb
Reviewed-on: https://chromium-review.googlesource.com/280959
Reviewed-by: Yuli Huang <yuli@chromium.org>
Commit-Queue: Hung-ying Tyan <tyanh@chromium.org>
Tested-by: Hung-ying Tyan <tyanh@chromium.org>