Show test image in fullscreen for DUTs without internal display.
The test image was shown on the right of the internal display in
fullscreen assuming that the internal display exists. This does
not work for DUTs without internal display. This commit makes
the test image shown on the first display available if internal
display is not present.

BUG=chromium:429932
TEST=tested with subsequent related commits

Change-Id: I56c7f1deebc84e492f11f157b8663d98234614ea
Reviewed-on: https://chromium-review.googlesource.com/280958
Tested-by: Hung-ying Tyan <tyanh@chromium.org>
Reviewed-by: Yuli Huang <yuli@chromium.org>
Commit-Queue: Hung-ying Tyan <tyanh@chromium.org>