[autotest] dedupe suite_scheduler bug without build name
We are still creating one bug per board when a control file
is missing for a suite in a build. This is because
the board name is part of the build name.

This CL removes the build name from the suite_scheduler bug's
search marker.

BUG=None
TEST=File a bug.

Change-Id: I8425ef9f417617a4c515b01a75b30cf3cc1e7e55
Reviewed-on: https://chromium-review.googlesource.com/283050
Tested-by: Fang Deng <fdeng@chromium.org>
Reviewed-by: Mungyung Ryu <mkryu@google.com>
Commit-Queue: Fang Deng <fdeng@chromium.org>