autotest (wifi): enable MaskedBSSID to run on a single-stream PHY
By default, _get_phy_for_frequency() ignores any PHY that doesn't
support dual-stream operation. This minimizes the possibility of
surprises in test results. For example, performance tests will
have hugely different results if run on single-stream PHY.

In the case of MaskedBSSID, however, this restriction puts us
in a situation where requirements are hard to satisfy. Specifically,
we'd require: two dual-stream PHYs that operate in the same band.
(We can't simply run multiple APs on a single PHY, because the APs
are deliberately configured to use the same BSSID. We'd rather not
run on multiple bands, because that takes us farther away from
the AP behavior this test is meant to emulate.)

To support MaskedBSSID on diverse hardware, we relax the requirements.
In particular, we allow the use of single-stream PHYs for MaskedBSSID.
Since MaskedBSSID tests correctness, but not performance, this should
not yield any surprising results.

To provide determinism, the single-stream PHY is chosen only after
more capable PHYs are already in use.

BUG=chromium:504529
TEST=network_WiFi_MaskedBSSID (logs showed dual-stream was used first)
TEST=network_WiFi_SimpleConnect.wifi_check5VHT80

Change-Id: I34433cf0a766a15a1528838481be514ce3bcfed9
Reviewed-on: https://chromium-review.googlesource.com/283621
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>