autotest (wifi): improve reliability of RoamWifiEndToEnd
When the RoamWifiEndToEnd starts, it sets up an AP, and checks if
the AP is present in Chrome's network list. The test does not,
however, initiate a scan. Thus, it is up to chance whether or not
a scan has completed by the time the test examines the network
list.

To fix this, this CL adds a scan before looking for the AP. (The
scan_for_networks() method initiates a scan, and then sleeps for
long enough for the scan to succeed.)

BUG=chromium:504527
TEST=network_WiFi_RoamEndToEnd (10 iterations)

Change-Id: I5190c43e83a64750d5909be6843d58b3b0843081
Reviewed-on: https://chromium-review.googlesource.com/283544
Trybot-Ready: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Bindu Mahadev <bmahadev@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>